chore: integrate re-clamp with collapse state
